+++ Assets/Scripts/Buildings/CustomBuildingData.cs
@@ -4,7 +4,7 @@ namespace bts {
   [CreateAssetMenu(menuName = "Buildings/Data/Custom", fileName = "CustomBuildingData")]
   public class CustomBuildingData : CustomPlacedObjectData {
     public GemstoneDictionary buildingCosts;
-    public int healthAmount;
+    public float healthAmount;
     public Mesh ghostMesh;
   }
 }
