+++ Assets/Scripts/HealthSystem/Health.cs
@@ -4,18 +4,18 @@ namespace bts {
   public class Health {
     public event Action OnReset = delegate { };
     public event Action OnValueChange = delegate { };
-    public int MaxHealth { get; private set; }
-    public int CurrentHealth { get; protected set; }
-    public float HealthNormalized => (float)CurrentHealth / MaxHealth;
+    public float MaxHealth { get; private set; }
+    public float CurrentHealth { get; protected set; }
+    public float HealthNormalized => CurrentHealth / MaxHealth;
     public bool HasFullHealth => CurrentHealth == MaxHealth;
     public bool HasNoHealth => CurrentHealth == 0;
 
-    public Health(int maxHealth) {
+    public Health(float maxHealth) {
       MaxHealth = maxHealth;
       CurrentHealth = maxHealth;
     }
 
-    public virtual void Damage(int amount) {
+    public virtual void Damage(float amount) {
       CurrentHealth -= amount;
       if (CurrentHealth < 0) {
         CurrentHealth = 0;
@@ -24,7 +24,7 @@ namespace bts {
       OnValueChange.Invoke();
     }
 
-    public virtual void Heal(int amount) {
+    public virtual void Heal(float amount) {
       CurrentHealth += amount;
       if (CurrentHealth > MaxHealth) {
         CurrentHealth = MaxHealth;
